Configuration Source Extraction part 1/4 - Extract DTO Interfaces (#197)
Convert all concrete DTO classes from cloud.eppo.ufc.dto to interfaces
in cloud.eppo.api.dto with nested Default implementation classes.

Converted classes:
- Allocation, BanditFlagVariation, Shard, TargetingCondition, TargetingRule, Split
- BanditAttributeCoefficients, BanditNumericAttributeCoefficients, BanditCategoricalAttributeCoefficients
- BanditCoefficients, BanditModelData, BanditParameters, BanditParametersResponse
- BanditReference, FlagConfig, FlagConfigResponse, Variation

The ufc.dto package now only contains the adapters subdirectory.
