+++ src/main/java/cloud/eppo/api/dto/BanditAttributeCoefficients.java
@@ -1,4 +1,4 @@
-package cloud.eppo.ufc.dto;
+package cloud.eppo.api.dto;
 
 import cloud.eppo.api.EppoValue;
