+++ src/main/java/cloud/eppo/api/dto/BanditReference.java
@@ -1,0 +1,54 @@
+package cloud.eppo.api.dto;
+
+import java.util.List;
+import java.util.Objects;
+
+public interface BanditReference {
+  String getModelVersion();
+
+  List<BanditFlagVariation> getFlagVariations();
+
+  class Default implements BanditReference {
+    private final String modelVersion;
+    private final List<BanditFlagVariation> flagVariations;
+
+    public Default(String modelVersion, List<BanditFlagVariation> flagVariations) {
+      this.modelVersion = modelVersion;
+      this.flagVariations = flagVariations;
+    }
+
+    @Override
+    public String toString() {
+      return "BanditReference{"
+          + "modelVersion='"
+          + modelVersion
+          + '\''
+          + ", flagVariations="
+          + flagVariations
+          + '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+      if (o == null || getClass() != o.getClass()) return false;
+      BanditReference that = (BanditReference) o;
+      return Objects.equals(modelVersion, that.getModelVersion())
+          && Objects.equals(flagVariations, that.getFlagVariations());
+    }
+
+    @Override
+    public int hashCode() {
+      return Objects.hash(modelVersion, flagVariations);
+    }
+
+    @Override
+    public String getModelVersion() {
+      return modelVersion;
+    }
+
+    @Override
+    public List<BanditFlagVariation> getFlagVariations() {
+      return flagVariations;
+    }
+  }
+}
