+++ src/main/java/cloud/eppo/api/dto/EppoValueType.java
@@ -1,4 +1,4 @@
-package cloud.eppo.ufc.dto;
+package cloud.eppo.api.dto;
 
 public enum EppoValueType {
   NULL,
